Use uint rather than ulong to identify nodes

+++ TimeGraphDatabase.Engine/StorageRecord.cs
@@ -1,3 +1,3 @@
 namespace TimeGraphDatabase.Engine;
 
-public readonly record struct StorageRecord(ulong Timestamp, ulong LhsId, ulong RhsId, ulong RelationshipId);
+public readonly record struct StorageRecord(ulong Timestamp, uint LhsId, uint RhsId, uint RelationshipId);
